Extract video file helpers from pages helper
Reduce clutter in `PagesHelper`. Move to new `VideoFilesHelper`.
